raisync: move event fetcher to the events module
